requirements.txt
nose
sphinx
confluent-kafka
pyyaml